feat: add relations between User and Restaurant models, add some btn

+++ routes/modules/restaurants.js
@@ -4,24 +4,27 @@ const router = express.Router()
 
 const Restaurants = require('../../models/restaurant')
 
-
-
-// 新增餐廳
-router.post('/', (req, res) => {
-    return Restaurants.create(req.body)
-        .then(() => res.redirect("/"))
-        .catch(err => console.log(err))
-})
-
 // 新增餐廳
 router.get('/new', (req, res) => {
     return res.render("new")
 })
 
+// 新增餐廳
+router.post('/', (req, res) => {
+    const userId = req.user._id
+
+    return Restaurants.create({ ...req.body, userId })
+        .then(() => res.redirect("/"))
+        .catch(err => console.log(err))
+})
+
+
 // 瀏覽特定餐廳
 router.get('/:restaurant_id', (req, res) => {
     const { restaurant_id } = req.params
-    return Restaurants.findById(restaurant_id)
+    const userId = req.user._id
+
+    return Restaurants.findOne({ _id: restaurant_id, userId })
         .lean()
         .then((restaurant) => res.render('show', { restaurant }))
         .catch((error) => console.log('error'))
@@ -30,7 +33,9 @@ router.get('/:restaurant_id', (req, res) => {
 // 編輯餐廳頁面
 router.get('/:restaurant_id/edit', (req, res) => {
     const { restaurant_id } = req.params
-    return Restaurants.findById(restaurant_id)
+    const userId = req.user._id
+
+    return Restaurants.findOne({ _id: restaurant_id, userId })
         .lean()
         .then(restaurant => res.render('edit', { restaurant }))
         .catch(err => console.log(err))
@@ -38,8 +43,10 @@ router.get('/:restaurant_id/edit', (req, res) => {
 
 // 更新餐廳
 router.put("/:restaurant_id", (req, res) => {
-    const { restaurant_id } = req.params;
-    return Restaurants.findByIdAndUpdate(restaurant_id, req.body)
+    const { restaurant_id } = req.params
+    const userId = req.user._id
+
+    return Restaurants.findByIdAndUpdate({_id: restaurant_id, userId}, req.body)
         .then(() => res.redirect(`/restaurants/${restaurant_id}`))
         .catch(error => console.log(error));
 })
@@ -48,7 +55,9 @@ router.put("/:restaurant_id", (req, res) => {
 // 刪除餐廳
 router.delete('/:restaurant_id', (req, res) => {
     const { restaurant_id } = req.params
-    return Restaurants.findByIdAndDelete(restaurant_id)
+    const userId = req.user._id
+
+    return Restaurants.findByIdAndDelete({ _id: restaurant_id, userId })
         .then(() => res.redirect('/'))
         .catch((err) => console.log('err'))
 })
